fix: Move provider to /src directory and small fixes
